pyproject.toml
[project]
name = "audio-qc-service"
version = "1.0.0"
description = "FastAPI Audio QC Service (VAD + metrics)"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn==0.34.2",
  "python-multipart==0.0.20",
  "numpy==1.26.4",
  "scipy==1.15.2",
  "soundfile==0.13.1",
  "funasr==1.2.6",
  "toml"
]

[tool.uvicorn]
# 这里只是占位，实际用命令行启动即可

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]
